Support function declarations and invocations
This translates function declarations into uninterpreted declarations,
and compiles invocations accordingly.  To make this work, declarations
must be stored in the environment so they can be recalled at the point
of invocation.

To correctly handle the translation of arguments, the two translation
methods (i.e. for `int` and for `bool`) have been combined into a single
`translate()` method which returns `Dynamic`.  That seems to be working
quite well.

Finally, note at this stage, checking preconditions at the point of
invocation is not yet supported.  Likewise, incorporating postconditions
is not supported either.