SMT solver now checks assertions
A process is spawned which fires up Z3 and pipes in the translated
SMT-LIB commands.  This is actually working, though with a number of
caveats:

1) The solver path is hard coded into the binary!

2) Assumes Z3 being used (i.e. specific command-line options are used).

3) No ability to set e.g. timeouts, etc.

The implementation using `std::process::Command` is not super efficient
or anything.  I'm still trying to figure out how you would use it
properly.  At this stage, it fires up a short-lived process and pipes in
the data, then gets the output.  It doesn't, for example, keep the
process alive or anything like that.